Updated LegacyCameraConnectionFragment so that will sense orientation of the device and rotate the camera.setDisplayOrientation(X) to appropriate angle so the display in the associated TextureView will be oriented correctly.
NOTE: have not corrected problem with needing to change the aspect ratio yet

NOTE 2:  Do not know if CameraConnectionFragment (used only if android 5.0 or more AND the hardware on device is FULL compliant --mine is not) will orient correctly as do not have device to test that is fully camera2 compliant but, from reading inherited code it should function correctly.   It should be noted that all emulator AVDs will run using LegacyCamera as it does not support hardware features of camera2 api (like control of focus etc).