Altered LegacyCameraConnectionFragment so it would alter aspect ration (wxh) versus (hxw) for changing device orientations between Landscape and Portrait.
NOTE: did a force test on using the camera2 api and the replacement class CameraConnectionFragment (instead of LegacyCameraConnectionFragment) and I believe that
class is working to.  For some reason the base inherited tf lite code ONLY allows use of camera 2 api if the hardware FULLY supports (there are different levels) the
camera 2 api.   This is odd because there is no real use of hardware camera 2 api like focusing...but, leaving as is.   However, in future old camera api may be removed
and this code base will need to be altered to simply always use camera 2 api.   See "usecamera2api" flag in code to understand where it is set and how it tests the hardware support.
Obviously all emulators will not support hardware features and hence usecamera2api will be false and the LegacyCameraConnectionFragment will be used.