Added overridden methods in FirebaseStorageUtil to be able to store images (if flag set) and the MaskRecord or FeverRecord or CrowdRecord or SocDistRecord associated with it
NOTE:  Cemil and Sahil --Even though I wrote the code I want you to be in charge of it.   So please read.

NOTE: NOW you are required to sign in as it tries to detect userid and creates iamge filenames based on it and timestamp.   NOTE...it may fail if not signed in ---we need to test this
and either fix apporpriately or ensure it will never happen.